Convert everything from Float16 to Float32.
Make it easier to convert back later if need be

+++ NeuralObscura/Globals.swift
@@ -9,4 +9,4 @@
 import Foundation
 import MetalPerformanceShaders
 
-let textureFormat = MPSImageFeatureChannelFormat.float16
+let textureFormat = MPSImageFeatureChannelFormat.float32
